Major rewrite
I decided to make ooida much simpler, and thus easier to test.
It still has all the features needed to parallelize it, especially the fill-in stage. But now there is only one command, the "ida" class, and a handful of methods. The ::ooida::HuntFill procedure is a procedural version of the IDA algorithm, and is the engine behind the object-oriented version.
I added a ton of tests, and cleaned up documentation.